requirements.txt
discord.py
json
